server/hash-worker/requirements.txt
psycopg2-binary==2.9.9
boto3==1.34.0
Pillow==10.2.0
imagehash==4.3.1
